scripts/pyproject.toml
[build-system]
requires = ["setuptools>=69"]
build-backend = "setuptools.build_meta"

[project]
name = "workspace-scripts"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "jinja2",
    "litellm",
    "pydantic",
    "pyyaml",
]

[tool.setuptools]
package-dir = {"" = ".."}

[tool.setuptools.packages.find]
where = [".."]
include = ["scripts*"]
